실습용/ai-assistant-project/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
easyocr
python-multipart
jinja2
python-dotenv
pandas
pydantic
pydantic-settings
requests
fastmcp
pdfplumber
python-docx
openpyxl
langchain-text-splitters
langchain-huggingface
